+++ actix-files/CHANGES.md
@@ -2,10 +2,11 @@
 
 ## Unreleased
 
+- Fix handling of special characters in filenames.
+
 ## 0.6.4
 
-- Fix handling of linebreaks in filenames. [#3237]
-- Fix handling of newlines in filenames. [#3235]
+- Fix handling of newlines in filenames.
 - Minimum supported Rust version (MSRV) is now 1.68 due to transitive `time` dependency.
 
 ## 0.6.3